move UnitTestDefinition, implement missing resource_class methods
